refactored error.go into own directory and updated instances to reflect change
