+++ customerrors/error.go
@@ -1,4 +1,4 @@
-package model
+package customerrors
 
 import (
 	"fmt"
@@ -6,21 +6,21 @@ import (
 	"os"
 )
 
-func fatalErrorHandler(err error) {
+func FatalErrorHandler(err error) {
 	//if there is an error with opening/finding the image, it will output a fatal error
 	if err != nil {
 		log.Fatal(err)
 	}
 }
 
-func exitErrorHandler(err error) {
+func ExitErrorHandler(err error) {
 	if err != nil {
 		fmt.Print(err.Error())
 		os.Exit(1)
 	}
 }
 
-func printErrorHandler(err error) {
+func PrintErrorHandler(err error) {
 	if err != nil {
 		fmt.Print(err.Error())
 	}
